fix: checklist unable to render
